Area for Improvements:- Change **"Develop a Spring Boot REST API"** to **"A Spring Boot API"** in the HLD. - Expand on **edge cases** in the relevant section. - Add a section for **Expected Behavior**. - Rewrite the **Plan** in **first-person future tense**. - Include the **Goal** section. - Format **relative paths** using backticks. - Fix **indentation** in the **How to Run** section. - Rename the title from **"Maxflow"** to **"Accessing Endpoints & Features"**. - Remove the **Security** section.
Iteration

On service layer
this entire code can be wrapped in a try catch. As this is a DSA question. sometimes null pointer may get thrown, which we may miss. So generally wrap everything with try catch for safer side and throw a MaxFlowException with some generic message, that something went wrong while running the algo